Add ssh and node listing commands, get corosync and pacemaker working.
Add command for ssh'ing into an individual node and a command that
gets the list of IPs for nodes and dumps the necessary config file
template (which is required until we have a version of corosync that
can add and remove nodes using udpu dynamically).

Also add real puppet manifests for corosync and pacemaker. They do all
the necessary setup with installing services, grabbing or generating
the necessary configs, and running the service. Unfortunately some
steps are still run manually, but much of the work is automated.