Add readiness indicators for Sirikata and Pacemaker.
Add files to the Puppet configuration which indicate that the Sirikata
and Pacemaker steps have all completed so the node should be usable
for these two things. Add a function for waiting on the the nodes to
become ready using a similar approach to waiting for them to be
pingable, but using an ssh command to test for the readiness indicator
files. Use this in addition to the ping wait when booting nodes so
that command can block until the nodes are completely ready.