Add ability to set node types and add redis nodes.
Add a command to set node types. The node type just adds to a config
file controlling what class is used for the node in Puppet. Also add a
Redis server module for Puppet and the appropriate node type that can
be inherited from when setting the type to sirikata_redis.