Add NodeGroupConfig and NodeGroup classes.
Refactor the existing ClusterConfigFile class into NodeGroupConfig, a
generic base class, and a subclass EC2GroupdConfig for EC2-specific
configuration. Also add the NodGroup class which will provide an API
for interacting with a group of nodes rather than just a command line
interface. The first implementation is for EC2 and currently they only
have a boot method.

Mostly this involves reorganizing code, making it generic, and fixing
assumptions that are no longer true. The biggest one is that we're not
just going to pass cluster names through the ec2 commands, instead
we'll support names (for the command line) and configs (for the API).