Add more commands to NodeGroups.
A few are just properties which give you default values so NodeGroups
with different configurations can be used together: user, sirikata
path, and default working path.

The rest are simple wrappers around the core commands for using the
cluster -- getting a list of nodes, adding and removing running
services, and terminating nodes.