Fix status checks on ec2.
There are various problems with the previous setup. We can still use
the approach, but we need to set a monitoring interval to even get
monitoring turned on, try to force failures to just cause the service
to be stopped, and clean out any old failures before starting so that
previous failures don't keep a service from starting up just because
it has the same name. Making failures trigger stops is actually broken
on some versions of Pacemaker, so we also set a migration-threshold,
which forces the service not to run on the current node after X
failures. Combined with location constraints, this stops the service
from running.