Switch to using a generic URL for Sirikata binaries.
As you scale up, the puppet master becomes a bottleneck. For some
reason it is very CPU intensive, and even if it weren't sometimes the
puppet master is not conveniently located w.r.t. the puppets. This
approach lets you do things like upload once to S3 and then have EC2
nodes pull from S3, which is very fast and free, as opposed to
uploading the same file into many different EC2 nodes.