Remove corosync and pacemaker support.
They were causing more trouble than they are currently worth. Long
term they would be nice as they allow for nice failover when not tying
services to specific nodes, but corosync would sometimes get into a
state where it would eat up a ton of CPU and bandwidth and never seem
to get back to a good state. Also, adding, checking the status of, and
removing services seemed to become increasingly unreliable as the
number of servers increased. We can revive the code later if it's
useful as a new starting point.

In the process of that change, this also added caching of some
information about instances so we don't have to hit the Amazon API
constantly just to get IP address.