Fix collection of reserved node info.
We were caching information, but some of it wasn't immediately
available for reserved instances, where we get the IDs immediately but
they are still getting assigned resources and being booted up. Add a
loop with sleeping to wait until that basic info (IPs and hostnames)
are available for all the instances.